+++ example/core/dashboard.py
@@ -2,7 +2,7 @@ from jet.dashboard import modules
 from jet.dashboard.dashboard import DefaultAppIndexDashboard
 
 
-class EventsStatsModule(modules.DashboardModule):
+class CustomModule(modules.DashboardModule):
     title = "Module"
     template = 'core/module.html'
 
@@ -15,8 +15,8 @@ class CustomAppIndexDashboard(DefaultAppIndexDashboard):
         for mod in modules.LinkList, modules.ModelList, modules.RecentActions:
             self.available_children.append(mod)
 
-        self.available_children.append(EventsStatsModule)
+        self.available_children.append(CustomModule)
         self.children[1].column = 2
         self.children.append(
-            EventsStatsModule(column=1, order=0)
+            CustomModule(column=1, order=0, collapsed=True)
         )